+++ NEWS.md
@@ -1,3 +1,9 @@
+# CHANGES IN knitr VERSION 1.2
+
+## MAJOR CHANGES
+
+- the document hook `hook_rjournal()` was removed; it was too hackish (see http://yihui.name/en/2013/02/contribute-to-the-r-journal-with-lyx-knitr/ for how to write an article for The R Journal in a less hackish way)
+
 # CHANGES IN knitr VERSION 1.1
 
 ## NEW FEATURES